Remove "enabled" property of resources.
In my opinion mods should remove it from all templates if they wish to.
Having these remains in the template isn't a clean approach to modding and
it shouldn't be hard to remove those by searching for the names of the affected resources
or just following the trail of errormessages of unidentified resources.

The enabled check implementation was incomplete as f.e. the structure tree and
Player component didn't check for these resources. Furthremore we don't want to
pay with performance penalties for this semifeature.