New survival of the fittest attacker wave formula.
Use exponential growth limited to 400 attackers, to guarantee ending the game in any case after 2 hours
while keeping the stength of the attackers in the first 30 minutes similar.

Distinguish between attacker types to support switching from citizen soldiers to champs later
and randomize the siege to champion ratio.

Move wave balancing constants to the top of the file.