Prevent NaN resource amounts if a non-natural batch train size was chosen
Summary:
As reported by user1 and some other lobby players, if a non-natural number is used in the batch-train-size option,
the simulation becomes broken in various ways.

Test Plan:
Don't apply the patch, start a game, open the options, set a string that isn't a number and observe
the GUI breaking when pressing shift and training something, as well as
having NaN resources, that avoid the resource cost restriction.
A non-natural number as a string will train N % 1 units and the remaining fraction can't be trained.
The population counter can also becomes a non-natural number.

Apply only the simulation patch to verify that even malicious multiplayer clients can't break the simulation anymore when controlling their GUI.
Apply the entire patch to verify that casual player can't break anything, even if they type a non-natural number or non-number in that option field.

Differential Revision: https://code.wildfiregames.com/D66

Trac Tickets: #4459